added check access token

+++ api.py
@@ -1,4 +1,6 @@
 import requests
+import os
+from datetime import datetime
 
 
 def get_access_token(client_id: str) -> str:
@@ -10,10 +12,21 @@ def get_access_token(client_id: str) -> str:
             }
         )
     response_token_request.raise_for_status()
-    return response_token_request.json()['access_token']
+    timestamp = float(datetime.now().timestamp())
+    os.environ.setdefault('ACCESS_TOKEN', response_token_request.json()['access_token'])
+    os.environ.setdefault('MOLTIN_TOKEN_EXPIRES_TIME', str(timestamp))
 
 
-def get_products(access_token: str) -> dict:
+def check_access_token(client_id: str):
+    token_expires_time = os.getenv('MOLTIN_TOKEN_EXPIRES_TIME')
+    timestamp = float(datetime.now().timestamp())
+    if not token_expires_time or float(token_expires_time) < timestamp:
+        get_access_token(client_id)
+
+
+def get_products(client_id: str) -> dict:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     response_products = requests.get(
         'https://api.moltin.com/v2/products',
@@ -23,7 +36,9 @@ def get_products(access_token: str) -> dict:
     return response_products.json()
 
 
-def get_product(access_token: str, id: str) -> dict:
+def get_product(client_id: str, id: str) -> dict:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     response_products = requests.get(
         f'https://api.moltin.com/v2/products/{id}',
@@ -33,7 +48,9 @@ def get_product(access_token: str, id: str) -> dict:
     return response_products.json()
 
 
-def get_image_product(access_token: str, id_image: str) -> dict:
+def get_image_product(client_id: str, id_image: str) -> dict:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     response_products = requests.get(
         f'https://api.moltin.com/v2/files/{id_image}',
@@ -43,7 +60,9 @@ def get_image_product(access_token: str, id_image: str) -> dict:
     return response_products.json()
 
 
-def add_product_cart(product: dict, access_token: str, quantity: int, cart_id: str) -> None:
+def add_product_cart(product: dict, client_id: str, quantity: int, cart_id: str) -> None:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     json_data = {
         'data': {
@@ -60,7 +79,9 @@ def add_product_cart(product: dict, access_token: str, quantity: int, cart_id: s
     response_add_product_to_cart.raise_for_status()
 
 
-def remove_product_from_cart(cart_id: str, product_id: str, access_token: str) -> None:
+def remove_product_from_cart(cart_id: str, product_id: str, client_id: str) -> None:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     response_add_product_to_cart = requests.delete(
         f'https://api.moltin.com/v2/carts/{cart_id}/items/{product_id}',
@@ -69,7 +90,9 @@ def remove_product_from_cart(cart_id: str, product_id: str, access_token: str) -
     response_add_product_to_cart.raise_for_status()
 
 
-def get_cart(cart_id: str, access_token: str) -> dict:
+def get_cart(cart_id: str, client_id: str) -> dict:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     response_get_cart = requests.get(
         f'https://api.moltin.com/v2/carts/{cart_id}/items',
@@ -79,7 +102,9 @@ def get_cart(cart_id: str, access_token: str) -> dict:
     return response_get_cart.json()
 
 
-def get_cart_total(cart_id: str, access_token: str) -> str:
+def get_cart_total(cart_id: str, client_id: str) -> str:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     response_get_cart = requests.get(
         f'https://api.moltin.com/v2/carts/{cart_id}',
@@ -91,7 +116,9 @@ def get_cart_total(cart_id: str, access_token: str) -> str:
     return total_cart
 
 
-def create_customer(name: str, email: str, access_token: str) -> None:
+def create_customer(name: str, email: str, client_id: str) -> None:
+    check_access_token(client_id)
+    access_token = os.getenv('ACCESS_TOKEN')
     headers = {'Authorization': access_token}
     json_data = {
         'data': {
